refactor(web): simplify header path structure
Moved 'header.tsx' to a higher directory level for a streamlined
import path. Updated related import statements in the header file to
reflect the new path structure. This change improves code readability
 and maintainability by reducing nested directory complexity.
