+++ apps/web/app/shell/header.tsx
@@ -6,10 +6,10 @@ import { SuddenlyGiovanni } from '@suddenlygiovanni/ui/components/suddenly-giova
 import { useToggle } from '@suddenlygiovanni/ui/hooks/use-toggle.tsx'
 import { clsx } from '@suddenlygiovanni/ui/lib/utils.ts'
 
-import avatarAssetUrl from '../../assets/giovanni_ravalico-profile_bw.webp'
-import { routesRecord } from '../../routes-record.ts'
-import { NavLink } from './nav-link.tsx'
-import { ThemeSwitch } from './theme-switch.tsx'
+import avatarAssetUrl from '../assets/giovanni_ravalico-profile_bw.webp'
+import { routesRecord } from '../routes-record.ts'
+import { NavLink } from './header/nav-link.tsx'
+import { ThemeSwitch } from './header/theme-switch.tsx'
 
 const routes = (
 	[
