+++ apps/web/app/shell/index.tsx
@@ -1,4 +1,4 @@
 export { Main } from './main.tsx'
-export { Header } from './header/header.tsx'
+export { Header } from './header.tsx'
 export { Footer } from './footer.tsx'
 export { Document } from './document.tsx'
